+++ quadratic-client/src/app/ui/components/LanguageIcon.tsx
@@ -10,15 +10,15 @@ interface LanguageIconProps extends SvgIconProps {
 
 export function LanguageIcon({ language, ...props }: LanguageIconProps) {
   return language === 'Python' ? (
-    <Python sx={{ color: colors.languagePython }} {...props} />
+    <Python {...props} sx={{ color: colors.languagePython, ...(props.sx ? props.sx : {}) }} />
   ) : language === 'Formula' ? (
-    <Formula sx={{ color: colors.languageFormula }} {...props} />
+    <Formula {...props} sx={{ color: colors.languageFormula, ...(props.sx ? props.sx : {}) }} />
   ) : language === 'Javascript' ? (
-    <JavaScript className="text-gray-700" sx={{ color: colors.languageJavascript }} />
+    <JavaScript className="text-gray-700" sx={{ color: colors.languageJavascript, ...(props.sx ? props.sx : {}) }} />
   ) : language === 'POSTGRES' ? (
-    <PostgresIcon sx={{ color: colors.languagePostgres }} {...props} />
+    <PostgresIcon {...props} sx={{ color: colors.languagePostgres, ...(props.sx ? props.sx : {}) }} />
   ) : language === 'MYSQL' ? (
-    <MysqlIcon sx={{ color: colors.languageMysql }} {...props} />
+    <MysqlIcon {...props} sx={{ color: colors.languageMysql, ...(props.sx ? props.sx : {}) }} />
   ) : (
     <Subject {...props} />
   );
